Refactoring TcpDataFlowExample
