+++ app/src/main/java/com/nio/example/event/EventHandler.java
@@ -1,0 +1,11 @@
+package com.nio.example.event;
+
+import java.io.IOException;
+import java.nio.channels.SelectionKey;
+
+public interface EventHandler {
+
+  boolean select(SelectionKey selectionKey);
+
+  void handle(SelectionKey selectionKey) throws IOException;
+}
